#Database + add repositories + fix repo config + refactor interface, model and entities + fix module-database
(cherry picked from commit 433e3a5e59bb26aa903caec48fd71b487ee13438)
