+++ packages/entities/src/Post.ts
@@ -1,15 +1,15 @@
-import {IPost, IAuthor} from "@mono/interfaces";
+import {Post} from '@mono/models';
 import {Column, Entity, ManyToOne, PrimaryGeneratedColumn} from 'typeorm'
 import {AuthorEntity} from "./Author";
 
 @Entity({name: "posts"})
-export class PostEntity implements IPost {
+export class PostEntity extends Post {
   @PrimaryGeneratedColumn("increment")
   id: string;
   @Column()
   content: string;
   @Column()
   title: string;
-  @ManyToOne(type => AuthorEntity, (author: IAuthor) => author.posts)
+  @ManyToOne(type => AuthorEntity, (author: AuthorEntity) => author.posts)
   author: AuthorEntity
 }
