+++ packages/interfaces/src/IAuthor.ts
@@ -1,7 +1,4 @@
-import {IPost} from "./IPost";
-
 export interface IAuthor {
   id: string
   name: string
-  posts:IPost[]
 }
