+++ packages/interfaces/src/IPost.ts
@@ -1,8 +1,5 @@
-import {IAuthor} from "./IAuthor";
-
 export interface IPost {
   id: string
   title: string
   content: string
-  author:IAuthor
 }
