+++ packages/models/src/Post.ts
@@ -1,5 +1,11 @@
 import {IPost} from "@mono/interfaces";
+import {Author} from "./Author";
 
-export class Post implements IPost{
+export abstract class Post implements IPost {
+
+  id: string;
+  title: string;
+  content: string;
+  author: Author;
 
 }
