+++ packages/models/src/index.ts
@@ -1,0 +1,6 @@
+/**
+ * @file Automatically generated by barrelsby.
+ */
+
+export * from "./Author";
+export * from "./Post";
