+++ packages/models/tsconfig.json
@@ -1,4 +1,12 @@
 {
   "extends": "../../tsconfig.json",
-  "exclude": ["node_modules", "dist"]
+  "exclude": [
+    "node_modules",
+    "dist"
+  ],
+  "references": [
+    {
+      "path": "../interfaces"
+    }
+  ]
 }
