+++ packages/module-database/package.json
@@ -11,7 +11,7 @@
     "generate-barrels": "barrelsby --delete  -d src -l top"
   },
   "dependencies": {
-    "@mono/interfaces": "0.1.0",
+    "@mono/entities": "0.1.0",
     "@nestjs/typeorm": "^6.1.3",
     "pg": "^7.12.1",
     "typeorm": "^0.2.18"
