+++ packages/repositories/index.d.js
@@ -1,0 +1,2 @@
+export * from './src';
+
